refactor: extract refactor-ops handlers from App.svelte (#670, increment 6)
Move the refactor/transform handler cluster out of App.svelte, following the
established create<Group>Ops(ctx) + feature-state-store + paired-test pattern.

- lib/stores/refactor-flow.svelte.ts — getRefactorFlowStore(): the two
  Auto-link review flows' in-flight state + the autoLinkBusy guard (was three
  App $state decls), read reactively by the handlers and the template's
  AutoLink / AutoLinkInbound dialogs.
- lib/app/refactor-ops.svelte.ts — createRefactorOps(ctx) houses 13 handlers
  moved VERBATIM: extract-selection, split-by-heading, split-here, auto-link
  (outbound/inbound + their apply), bulk add/remove tag, toggle-entrypoint,
  format, bibliography, auto-tag (plus internal resolveTitle / bulkTagTargets /
  reportBulkTagSummary). It's a `.svelte.ts` module because the auto-link apply
  handlers use the `$state.snapshot` rune. Substitutions only: editorComponent
  / sidebar via ctx getters, maybeHandleMissingApiKey via ctx (it drives App
  settings-dialog state so it stays in App), autoLink state via the flow store,
  busy via the busy store.
- App builds the ctx, destructures the handlers back into the same names; the
  AutoLink dialogs now read refactorFlow.*. Tool-routing (handleDecompose /
  handleCrystallize) and the missing-API-key flow stay in App.
- tests/renderer/app/refactor-ops.test.ts (11 tests): auto-tag success/empty/
  missing-key, auto-link review-set/none + apply-clears, add-tag no-targets vs
  write, bibliography no-tab vs generate.

App.svelte: 2,612 → 2,083 (cumulative 3,764 → 2,083, −1,681).

Verified: pnpm lint (0 errors), full suite (2813 passed), e2e boot smoke.
Reactivity reviewed by hand (runes store; ctx getters; reactive template reads).
